refactor(cli): family-specific tests を family modules へ relocate (SPEC-1942)
SPEC-1942 SC-025 (cli.rs <1000 行) 達成。 共有 fake_gh 基盤 (compile_fake_gh
/ with_fake_gh / fake_gh_test_lock / sample_thread / sample_issue_snapshot
/ sample_pr_status / ScopedEnvVar / commands_for_event) を新規
cli/test_support.rs (cfg(test) only) に集約。 各 family-specific test を
該当 family module の tests へ relocate。

cli/pr.rs::tests へ relocate (7 tests):
- review_thread_reply_is_skipped_for_duplicate_body
- review_thread_reply_is_skipped_for_resolved_or_outdated_threads
- pr_checks_response_returns_error_when_gh_fails
- pr_checks_response_parses_success_payload
- gh_wrappers_parse_successful_responses
- gh_wrappers_cover_none_fallback_and_zip_error_paths
- gh_wrappers_tolerate_resolve_failure_after_remote_state_changes

cli/issue.rs::tests へ relocate (1 test):
- cache_backed_issue_and_linked_pr_helpers_reuse_cached_data

cli/hook/mod.rs::tests へ relocate (2 tests):
- daemon_hook_argv_and_internal_command_output_preserve_streams
- hook_front_door_refresh_migrates_stale_multi_hook_settings

cli.rs に残置 (4 tests): cli core / family split RED test / cross-family
render テスト群:
- parse_and_guard_helpers_cover_additional_error_paths
- cli_command_family_split_round_trip_parses
- render_helpers_include_issue_pr_and_review_details
- render_helpers_cover_empty_states_and_url_parsing_fallbacks
- cache_and_parse_helpers_cover_fallback_paths

cli.rs 1545 → **920 行 (-625)**、 累計 -1781 from 2701 開始時 (-65.9%)、
SC-025 (<1000 行) 達成。 機能・契約・test 結果は不変 (lib 345 + integration
全 pass、 clippy 0 / fmt clean)。
