+++ crates/gwt/src/cli.rs
@@ -31,6 +31,8 @@ mod issue_spec;
 mod plan;
 mod pr;
 mod skill_state_runtime;
+#[cfg(test)]
+mod test_support;
 pub mod update;
 
 use std::io::{self};
@@ -593,325 +595,17 @@ pub(crate) fn fake_gh_test_lock() -> &'static std::sync::Mutex<()> {
 
 #[cfg(test)]
 mod tests {
-    use std::{
-        env, fs,
-        path::{Path, PathBuf},
-    };
-
     use tempfile::tempdir;
 
-    use crate::cli::env::{InternalCommandOutput, TestEnv};
-    use gwt_github::{
-        CommentId, CommentSnapshot, IssueNumber, IssueSnapshot, IssueState, UpdatedAt,
-    };
+    use gwt_github::{IssueNumber, IssueSnapshot, IssueState};
 
     use super::*;
-
-    fn compile_fake_gh(bin_dir: &Path) {
-        let source = r###"
-use std::{env, fs, process::ExitCode};
-
-fn pr_json(number: &str, title: &str) -> String {
-    format!(
-        "{{\"number\":{number},\"title\":\"{title}\",\"state\":\"OPEN\",\"url\":\"https://github.com/akiojin/gwt/pull/{number}\",\"mergeable\":\"MERGEABLE\",\"mergeStateStatus\":\"CLEAN\",\"statusCheckRollup\":[{{\"name\":\"ci\",\"status\":\"COMPLETED\",\"conclusion\":\"SUCCESS\"}}],\"reviewDecision\":\"APPROVED\"}}"
-    )
-}
-
-fn behind_pr_json(number: &str, title: &str) -> String {
-    format!(
-        "{{\"number\":{number},\"title\":\"{title}\",\"state\":\"OPEN\",\"url\":\"https://github.com/akiojin/gwt/pull/{number}\",\"mergeable\":\"MERGEABLE\",\"mergeStateStatus\":\"BEHIND\",\"statusCheckRollup\":[{{\"name\":\"ci\",\"status\":\"COMPLETED\",\"conclusion\":\"SUCCESS\"}}],\"reviewDecision\":\"REVIEW_REQUIRED\"}}"
-    )
-}
-
-fn review_threads_json(resolved_after_fail: bool) -> String {
-    let resolved = if resolved_after_fail { "true" } else { "false" };
-    r#"{"data":{"repository":{"pullRequest":{"reviewThreads":{"nodes":[
-{"id":"thread-1","isResolved":__RESOLVED__,"isOutdated":false,"path":"src/lib.rs","line":10,"comments":{"nodes":[{"id":"comment-1","body":"done","createdAt":"2026-04-20T00:00:00Z","updatedAt":"2026-04-20T00:00:00Z","author":{"login":"reviewer"}}]}},
-{"id":"thread-2","isResolved":false,"isOutdated":false,"path":"src/main.rs","line":12,"comments":{"nodes":[{"id":"comment-2","body":"needs changes","createdAt":"2026-04-20T01:00:00Z","updatedAt":"2026-04-20T01:00:00Z","author":{"login":"reviewer"}}]}}
-]}}}}}"#
-        .replace("__RESOLVED__", resolved)
-}
-
-fn main() -> ExitCode {
-    let args: Vec<String> = env::args().skip(1).collect();
-    let mode = env::var("GWT_FAKE_GH_MODE").unwrap_or_else(|_| "success".to_string());
-    let state_file = env::var("GWT_FAKE_GH_STATE_FILE").ok();
-
-    match args.as_slice() {
-        [pr, view, json_flag, ..] if pr == "pr" && view == "view" && json_flag == "--json" => {
-            if mode == "no-current-pr" {
-                eprintln!("no pull requests found for branch");
-                return ExitCode::from(1);
-            }
-            if mode == "behind" {
-                println!("{}", behind_pr_json("12", "Current PR"));
-            } else {
-                println!("{}", pr_json("12", "Current PR"));
-            }
-            return ExitCode::SUCCESS;
-        }
-        [pr, view, number, repo_flag, _, json_flag, ..]
-            if pr == "pr" && view == "view" && repo_flag == "--repo" && json_flag == "--json" =>
-        {
-            if mode == "behind" {
-                println!("{}", behind_pr_json(number, "Fetched PR"));
-            } else {
-                println!("{}", pr_json(number, "Fetched PR"));
-            }
-            return ExitCode::SUCCESS;
-        }
-        [pr, create, ..] if pr == "pr" && create == "create" => {
-            println!("https://github.com/akiojin/gwt/pull/12");
-            return ExitCode::SUCCESS;
-        }
-        [pr, edit, ..] if pr == "pr" && edit == "edit" => {
-            return ExitCode::SUCCESS;
-        }
-        [pr, comment, ..] if pr == "pr" && comment == "comment" => {
-            return ExitCode::SUCCESS;
-        }
-        [pr, checks, _, json_flag, fields] if pr == "pr" && checks == "checks" && json_flag == "--json" => {
-            if mode == "checks-fallback" && !fields.contains("bucket") {
-                eprintln!("unknown JSON field\nAvailable fields:\n  name\n  state\n  bucket\n  link\n  startedAt\n  completedAt\n  workflow");
-                return ExitCode::from(1);
-            }
-            if fields.contains("bucket") {
-                println!("[{{\"name\":\"CI\",\"state\":\"COMPLETED\",\"bucket\":\"pass\",\"link\":\"https://example.test/checks/12\",\"startedAt\":\"2026-04-20T00:00:00Z\",\"completedAt\":\"2026-04-20T00:01:00Z\",\"workflow\":\"coverage\"}}]");
-            } else {
-                println!("[{{\"name\":\"CI\",\"state\":\"COMPLETED\",\"conclusion\":\"SUCCESS\",\"detailsUrl\":\"https://example.test/checks/12\",\"startedAt\":\"2026-04-20T00:00:00Z\",\"completedAt\":\"2026-04-20T00:01:00Z\"}}]");
-            }
-            return ExitCode::SUCCESS;
-        }
-        [run, view, run_id, log_flag] if run == "run" && view == "view" && log_flag == "--log" => {
-            println!("run log {run_id}");
-            return ExitCode::SUCCESS;
-        }
-        [api, endpoint] if api == "api" && endpoint == "repos/akiojin/gwt/pulls/12/reviews" => {
-            println!("[{{\"id\":42,\"state\":\"APPROVED\",\"body\":\"Looks good\",\"submitted_at\":\"2026-04-20T02:00:00Z\",\"user\":{{\"login\":\"reviewer\"}}}}]");
-            return ExitCode::SUCCESS;
-        }
-        [api, endpoint] if api == "api" && endpoint == "/repos/akiojin/gwt/actions/jobs/91/logs" => {
-            if mode == "job-log-zip" {
-                print!("PKZIP");
-            } else {
-                print!("job log 91");
-            }
-            return ExitCode::SUCCESS;
-        }
-        [api, graphql, ..] if api == "api" && graphql == "graphql" => {
-            let joined = args.join("\n");
-            if joined.contains("timelineItems") {
-                println!(
-                    "{}",
-                    r#"{"data":{"repository":{"issue":{"timelineItems":{"nodes":[
-{"__typename":"CrossReferencedEvent","source":{"__typename":"PullRequest","number":12,"title":"Coverage Gate","state":"OPEN","url":"https://github.com/akiojin/gwt/pull/12"}},
-{"__typename":"ConnectedEvent","subject":{"__typename":"PullRequest","number":13,"title":"Follow-up","state":"MERGED","url":"https://github.com/akiojin/gwt/pull/13"}},
-{"__typename":"ConnectedEvent","subject":{"__typename":"PullRequest","number":12,"title":"Duplicate","state":"OPEN","url":"https://github.com/akiojin/gwt/pull/12"}}
-]}}}}}"#
-                );
-                return ExitCode::SUCCESS;
-            }
-            if joined.contains("reviewThreads") {
-                let resolved_after_fail = state_file
-                    .as_deref()
-                    .map(fs::metadata)
-                    .transpose()
-                    .ok()
-                    .flatten()
-                    .is_some();
-                println!("{}", review_threads_json(resolved_after_fail));
-                return ExitCode::SUCCESS;
-            }
-            if joined.contains("addPullRequestReviewThreadReply") {
-                println!("{{\"data\":{{\"addPullRequestReviewThreadReply\":{{\"comment\":{{\"id\":\"reply-1\"}}}}}}}}");
-                return ExitCode::SUCCESS;
-            }
-            if joined.contains("resolveReviewThread") {
-                if mode == "resolve-fails-but-resolved" {
-                    let already_failed = state_file
-                        .as_deref()
-                        .map(fs::metadata)
-                        .transpose()
-                        .ok()
-                        .flatten()
-                        .is_some();
-                    if !already_failed {
-                        if let Some(state_file) = state_file.as_deref() {
-                            let _ = fs::write(state_file, "resolved");
-                        }
-                        eprintln!("thread already resolved");
-                        return ExitCode::from(1);
-                    }
-                }
-                println!("{{\"data\":{{\"resolveReviewThread\":{{\"thread\":{{\"id\":\"thread-1\",\"isResolved\":true}}}}}}}}");
-                return ExitCode::SUCCESS;
-            }
-        }
-        _ => {}
-    }
-
-    eprintln!("unexpected fake gh args: {args:?}");
-    ExitCode::from(1)
-}
-"###;
-
-        let source_path = bin_dir.join("gh.rs");
-        fs::write(&source_path, source).expect("write fake gh source");
-        let output_path = bin_dir.join(format!("gh{}", env::consts::EXE_SUFFIX));
-        let status = std::process::Command::new("rustc")
-            .arg(&source_path)
-            .arg("-o")
-            .arg(&output_path)
-            .status()
-            .expect("compile fake gh");
-        assert!(status.success(), "fake gh compilation failed");
-    }
-
-    fn with_fake_gh<T>(mode: &str, test: impl FnOnce(&Path) -> T) -> T {
-        let _lock = super::fake_gh_test_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        let temp = tempdir().expect("tempdir");
-        compile_fake_gh(temp.path());
-
-        let repo_path = temp.path().join("repo");
-        fs::create_dir_all(&repo_path).expect("create repo path");
-
-        let old_path = env::var_os("PATH");
-        let old_mode = env::var_os("GWT_FAKE_GH_MODE");
-        let old_state = env::var_os("GWT_FAKE_GH_STATE_FILE");
-        let state_file = temp.path().join("gh-state");
-        let joined_path = env::join_paths(
-            std::iter::once(PathBuf::from(temp.path()))
-                .chain(old_path.iter().flat_map(env::split_paths)),
-        )
-        .expect("join PATH");
-        env::set_var("PATH", joined_path);
-        env::set_var("GWT_FAKE_GH_MODE", mode);
-        env::set_var("GWT_FAKE_GH_STATE_FILE", &state_file);
-
-        let result = test(&repo_path);
-
-        match old_path {
-            Some(value) => env::set_var("PATH", value),
-            None => env::remove_var("PATH"),
-        }
-        match old_mode {
-            Some(value) => env::set_var("GWT_FAKE_GH_MODE", value),
-            None => env::remove_var("GWT_FAKE_GH_MODE"),
-        }
-        match old_state {
-            Some(value) => env::set_var("GWT_FAKE_GH_STATE_FILE", value),
-            None => env::remove_var("GWT_FAKE_GH_STATE_FILE"),
-        }
-
-        result
-    }
-
-    fn sample_thread() -> PrReviewThread {
-        PrReviewThread {
-            id: "thread-1".to_string(),
-            is_resolved: false,
-            is_outdated: false,
-            path: "src/lib.rs".to_string(),
-            line: Some(12),
-            comments: vec![],
-        }
-    }
-
-    #[test]
-    fn review_thread_reply_is_skipped_for_duplicate_body() {
-        let mut thread = sample_thread();
-        thread.comments.push(PrReviewThreadComment {
-            id: "comment-1".to_string(),
-            body: "Fixed in latest commit.".to_string(),
-            created_at: "2026-04-10T00:00:00Z".to_string(),
-            updated_at: "2026-04-10T00:00:00Z".to_string(),
-            author: "reviewer".to_string(),
-        });
-
-        assert!(!crate::cli::pr::should_reply_to_review_thread(
-            &thread,
-            "Fixed in latest commit."
-        ));
-        assert!(crate::cli::pr::should_resolve_review_thread(&thread));
-    }
-
-    #[test]
-    fn review_thread_reply_is_skipped_for_resolved_or_outdated_threads() {
-        let mut resolved = sample_thread();
-        resolved.is_resolved = true;
-        assert!(!crate::cli::pr::should_reply_to_review_thread(
-            &resolved, "reply"
-        ));
-        assert!(!crate::cli::pr::should_resolve_review_thread(&resolved));
-
-        let mut outdated = sample_thread();
-        outdated.is_outdated = true;
-        assert!(!crate::cli::pr::should_reply_to_review_thread(
-            &outdated, "reply"
-        ));
-        assert!(!crate::cli::pr::should_resolve_review_thread(&outdated));
-    }
-
-    #[test]
-    fn pr_checks_response_returns_error_when_gh_fails() {
-        let err =
-            crate::cli::pr::parse_pr_checks_items_response("", "auth failed", false).unwrap_err();
-        assert!(
-            err.to_string().contains("gh pr checks: auth failed"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[test]
-    fn pr_checks_response_parses_success_payload() {
-        let items = crate::cli::pr::parse_pr_checks_items_response(
-            r#"[{"name":"test","state":"COMPLETED","conclusion":"SUCCESS","detailsUrl":"https://example.com","startedAt":"2026-04-10T00:00:00Z","completedAt":"2026-04-10T00:01:00Z","workflow":"CI"}]"#,
-            "",
-            true,
-        )
-        .unwrap();
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].name, "test");
-        assert_eq!(items[0].conclusion, "SUCCESS");
-    }
-
-    fn sample_issue_snapshot() -> IssueSnapshot {
-        IssueSnapshot {
-            number: IssueNumber(42),
-            title: "Coverage gate".to_string(),
-            body: "Raise the project coverage gate.\n".to_string(),
-            labels: vec!["gwt-spec".to_string(), "coverage".to_string()],
-            state: IssueState::Open,
-            updated_at: UpdatedAt::new("2026-04-20T00:00:00Z"),
-            comments: vec![CommentSnapshot {
-                id: CommentId(7),
-                body: "Need more tests.".to_string(),
-                updated_at: UpdatedAt::new("2026-04-20T01:00:00Z"),
-            }],
-        }
-    }
-
-    fn sample_pr_status() -> gwt_git::PrStatus {
-        gwt_git::PrStatus {
-            number: 128,
-            title: "Enforce coverage".to_string(),
-            state: gwt_git::pr_status::PrState::Open,
-            url: "https://github.com/akiojin/gwt/pull/128".to_string(),
-            ci_status: "SUCCESS".to_string(),
-            mergeable: "MERGEABLE".to_string(),
-            merge_state_status: "CLEAN".to_string(),
-            review_status: "APPROVED".to_string(),
-        }
-    }
 
     #[test]
     fn render_helpers_include_issue_pr_and_review_details() {
         let mut out = String::new();
-        let issue = sample_issue_snapshot();
-        let pr = sample_pr_status();
+        let issue = crate::cli::test_support::sample_issue_snapshot();
+        let pr = crate::cli::test_support::sample_pr_status();
 
         crate::cli::issue::render_issue(&mut out, &issue);
         crate::cli::issue::render_issue_comments(&mut out, &issue);
@@ -963,7 +657,7 @@ fn main() -> ExitCode {
                     updated_at: "2026-04-20T03:00:00Z".to_string(),
                     author: "reviewer".to_string(),
                 }],
-                ..sample_thread()
+                ..crate::cli::test_support::sample_thread()
             }],
         );
 
@@ -1027,122 +721,6 @@ fn main() -> ExitCode {
     }
 
     #[test]
-    fn daemon_hook_argv_and_internal_command_output_preserve_streams() {
-        let argv = crate::cli::hook::daemon_hook_argv(
-            "runtime-state",
-            &["start".to_string(), "--json".to_string()],
-        );
-        assert_eq!(
-            argv,
-            vec![
-                "gwtd".to_string(),
-                "__internal".to_string(),
-                "daemon-hook".to_string(),
-                "runtime-state".to_string(),
-                "start".to_string(),
-                "--json".to_string(),
-            ]
-        );
-
-        let temp = tempdir().expect("tempdir");
-        let mut env = TestEnv::new(temp.path().to_path_buf());
-        let status = crate::cli::hook::write_internal_command_output(
-            &mut env,
-            InternalCommandOutput {
-                status: 7,
-                stdout: b"stdout-bytes".to_vec(),
-                stderr: b"stderr-bytes".to_vec(),
-            },
-        )
-        .unwrap();
-
-        assert_eq!(status, 7);
-        assert_eq!(env.stdout, b"stdout-bytes");
-        assert_eq!(env.stderr, b"stderr-bytes");
-    }
-
-    #[test]
-    fn hook_front_door_refresh_migrates_stale_multi_hook_settings() {
-        let _env_lock = crate::env_test_lock()
-            .lock()
-            .unwrap_or_else(|poisoned| poisoned.into_inner());
-        let temp = tempdir().expect("tempdir");
-        let status = gwt_core::process::hidden_command("git")
-            .arg("init")
-            .arg("-q")
-            .current_dir(temp.path())
-            .status()
-            .expect("git init");
-        assert!(status.success(), "git init failed");
-
-        let hook_bin = temp.path().join("bin/gwtd");
-        fs::create_dir_all(hook_bin.parent().unwrap()).expect("create bin dir");
-        fs::write(&hook_bin, "#!/bin/sh\n").expect("write hook bin");
-        let _hook_bin = ScopedEnvVar::set("GWT_HOOK_BIN", &hook_bin);
-
-        let settings_path = temp.path().join(".claude/settings.local.json");
-        fs::create_dir_all(settings_path.parent().unwrap()).expect("create .claude");
-        fs::write(
-            &settings_path,
-            serde_json::to_string_pretty(&serde_json::json!({
-                "hooks": {
-                    "PreToolUse": [
-                        {
-                            "matcher": "*",
-                            "hooks": [
-                                {
-                                    "command": "'/tmp/bunx-old/gwtd' hook runtime-state PreToolUse",
-                                    "type": "command"
-                                }
-                            ]
-                        },
-                        {
-                            "matcher": "*",
-                            "hooks": [
-                                {
-                                    "command": "'/tmp/bunx-old/gwtd' hook forward",
-                                    "type": "command"
-                                }
-                            ]
-                        },
-                        {
-                            "matcher": "*",
-                            "hooks": [
-                                {
-                                    "command": "'/tmp/bunx-old/gwtd' hook workflow-policy",
-                                    "type": "command"
-                                }
-                            ]
-                        }
-                    ]
-                }
-            }))
-            .unwrap(),
-        )
-        .expect("write stale settings");
-
-        crate::cli::hook::refresh_managed_assets_for_hook_front_door(temp.path())
-            .expect("refresh hook assets");
-
-        let content = fs::read_to_string(&settings_path).expect("read settings");
-        let value: serde_json::Value = serde_json::from_str(&content).expect("settings json");
-        let commands = commands_for_event(&value, "PreToolUse");
-        assert_eq!(
-            commands.len(),
-            1,
-            "stale split hook entries must collapse to one dispatcher: {commands:?}"
-        );
-        assert!(
-            commands[0].contains(" hook event PreToolUse"),
-            "dispatcher command expected, got: {commands:?}"
-        );
-        assert!(
-            !content.contains("/tmp/bunx-old"),
-            "stale temporary hook binary path must be removed, got: {content}"
-        );
-    }
-
-    #[test]
     fn parse_and_guard_helpers_cover_additional_error_paths() {
         assert!(should_dispatch_cli(&[
             "gwt".to_string(),
@@ -1176,44 +754,11 @@ fn main() -> ExitCode {
         assert!(crate::cli::pr::edit_or_create_repo_guard("akiojin", "gwt").is_ok());
     }
 
-    struct ScopedEnvVar {
-        key: &'static str,
-        previous: Option<std::ffi::OsString>,
-    }
-
-    impl ScopedEnvVar {
-        fn set(key: &'static str, value: impl AsRef<std::ffi::OsStr>) -> Self {
-            let previous = std::env::var_os(key);
-            std::env::set_var(key, value);
-            Self { key, previous }
-        }
-    }
-
-    impl Drop for ScopedEnvVar {
-        fn drop(&mut self) {
-            if let Some(previous) = self.previous.as_ref() {
-                std::env::set_var(self.key, previous);
-            } else {
-                std::env::remove_var(self.key);
-            }
-        }
-    }
-
-    fn commands_for_event<'a>(value: &'a serde_json::Value, event: &str) -> Vec<&'a str> {
-        value["hooks"][event]
-            .as_array()
-            .unwrap_or_else(|| panic!("hooks missing for event {event}"))
-            .iter()
-            .flat_map(|entry| entry["hooks"].as_array().into_iter().flatten())
-            .filter_map(|hook| hook["command"].as_str())
-            .collect()
-    }
-
     #[test]
     fn render_helpers_cover_empty_states_and_url_parsing_fallbacks() {
         let issue = IssueSnapshot {
             comments: Vec::new(),
-            ..sample_issue_snapshot()
+            ..crate::cli::test_support::sample_issue_snapshot()
         };
         let mut out = String::new();
 
@@ -1251,176 +796,6 @@ fn main() -> ExitCode {
             ),
             vec!["number".to_string(), "title".to_string()]
         );
-    }
-
-    #[test]
-    fn cache_backed_issue_and_linked_pr_helpers_reuse_cached_data() {
-        let temp = tempdir().expect("tempdir");
-        let mut env = TestEnv::new(temp.path().to_path_buf());
-        let snapshot = sample_issue_snapshot();
-        env.client.seed(snapshot.clone());
-
-        let loaded = crate::cli::issue::load_or_refresh_issue(&mut env, snapshot.number, false)
-            .expect("load issue");
-        assert_eq!(loaded.snapshot.number, snapshot.number);
-        assert_eq!(env.client.call_log(), vec!["fetch:#42".to_string()]);
-
-        let cached = crate::cli::issue::load_or_refresh_issue(&mut env, snapshot.number, false)
-            .expect("cached issue");
-        assert_eq!(cached.snapshot.title, snapshot.title);
-        assert_eq!(env.client.call_log(), vec!["fetch:#42".to_string()]);
-
-        env.seed_linked_prs(
-            42,
-            vec![LinkedPrSummary {
-                number: 128,
-                title: "Enforce coverage".to_string(),
-                state: "OPEN".to_string(),
-                url: "https://github.com/akiojin/gwt/pull/128".to_string(),
-            }],
-        );
-        let linked =
-            crate::cli::issue::load_or_refresh_linked_prs(&mut env, snapshot.number, false)
-                .expect("linked prs");
-        assert_eq!(linked.len(), 1);
-        assert_eq!(env.linked_pr_calls(), vec![42]);
-
-        env.clear_linked_pr_calls();
-        let cached_linked =
-            crate::cli::issue::load_or_refresh_linked_prs(&mut env, snapshot.number, false)
-                .expect("cached linked prs");
-        assert_eq!(cached_linked.len(), 1);
-        assert!(env.linked_pr_calls().is_empty());
-
-        let cache_path = crate::cli::issue::linked_prs_cache_path(temp.path(), snapshot.number);
-        std::fs::create_dir_all(cache_path.parent().expect("cache dir")).expect("create cache dir");
-        std::fs::write(&cache_path, "{not-json").expect("write invalid json");
-        assert!(crate::cli::issue::read_linked_prs_cache(temp.path(), snapshot.number).is_err());
-    }
-
-    #[test]
-    fn gh_wrappers_parse_successful_responses() {
-        with_fake_gh("success", |repo_path| {
-            let linked =
-                crate::cli::issue::fetch_linked_prs_via_gh("akiojin", "gwt", IssueNumber(42))
-                    .expect("linked");
-            assert_eq!(linked.len(), 2);
-            assert_eq!(linked[0].number, 12);
-            assert_eq!(linked[1].state, "MERGED");
-
-            let current = crate::cli::pr::fetch_current_pr_via_gh(repo_path)
-                .expect("current pr")
-                .expect("current pr exists");
-            assert_eq!(current.number, 12);
-            assert_eq!(current.merge_state_status, "CLEAN");
-
-            let created = crate::cli::pr::create_pr_via_gh(
-                "akiojin/gwt",
-                repo_path,
-                &PrCreateCall {
-                    base: "develop".to_string(),
-                    head: Some("feature/coverage".to_string()),
-                    title: "Raise coverage".to_string(),
-                    body: "Body".to_string(),
-                    labels: vec!["coverage".to_string()],
-                    draft: true,
-                },
-            )
-            .expect("create pr");
-            assert_eq!(created.number, 12);
-
-            let edited = crate::cli::pr::edit_pr_via_gh(
-                "akiojin/gwt",
-                repo_path,
-                12,
-                Some("Edited"),
-                Some("Updated body"),
-                &["tested".to_string()],
-            )
-            .expect("edit pr");
-            assert_eq!(edited.number, 12);
-
-            crate::cli::pr::comment_on_pr_via_gh(repo_path, 12, "done").expect("comment");
-
-            let reviews =
-                crate::cli::pr::fetch_pr_reviews_via_gh("akiojin", "gwt", 12).expect("reviews");
-            assert_eq!(reviews.len(), 1);
-            assert_eq!(reviews[0].author, "reviewer");
-
-            let threads = crate::cli::pr::fetch_pr_review_threads_via_gh("akiojin", "gwt", 12)
-                .expect("review threads");
-            assert_eq!(threads.len(), 2);
-            assert_eq!(threads[0].line, Some(10));
-
-            let resolved = crate::cli::pr::reply_and_resolve_pr_review_threads_via_gh(
-                "akiojin", "gwt", 12, "done",
-            )
-            .expect("reply and resolve");
-            assert_eq!(resolved, 2);
-
-            let checks = crate::cli::pr::fetch_pr_checks_via_gh("akiojin/gwt", repo_path, 12)
-                .expect("checks");
-            assert!(checks.summary.contains("PR #12"));
-            assert_eq!(checks.checks.len(), 1);
-            assert_eq!(checks.checks[0].conclusion, "SUCCESS");
-
-            let run_log =
-                crate::cli::actions::fetch_actions_run_log_via_gh(repo_path, 90).expect("run log");
-            assert_eq!(run_log.trim(), "run log 90");
-
-            let job_log =
-                crate::cli::actions::fetch_actions_job_log_via_gh("akiojin", "gwt", repo_path, 91)
-                    .expect("job log");
-            assert_eq!(job_log, "job log 91");
-        });
-    }
-
-    #[test]
-    fn gh_wrappers_cover_none_fallback_and_zip_error_paths() {
-        with_fake_gh("no-current-pr", |repo_path| {
-            assert!(crate::cli::pr::fetch_current_pr_via_gh(repo_path)
-                .expect("current pr result")
-                .is_none());
-        });
-
-        with_fake_gh("checks-fallback", |repo_path| {
-            let checks = crate::cli::pr::fetch_pr_checks_via_gh("akiojin/gwt", repo_path, 12)
-                .expect("checks");
-            assert_eq!(checks.checks.len(), 1);
-            assert_eq!(checks.checks[0].workflow, "coverage");
-            assert_eq!(checks.checks[0].url, "https://example.test/checks/12");
-        });
-
-        with_fake_gh("behind", |repo_path| {
-            let current = crate::cli::pr::fetch_current_pr_via_gh(repo_path)
-                .expect("current pr")
-                .expect("current pr exists");
-            assert_eq!(current.effective_merge_status(), "BEHIND");
-
-            let checks = crate::cli::pr::fetch_pr_checks_via_gh("akiojin/gwt", repo_path, 12)
-                .expect("checks");
-            assert!(checks.summary.contains("Merge: BEHIND"));
-            assert_eq!(checks.merge_status, "BEHIND");
-        });
-
-        with_fake_gh("job-log-zip", |repo_path| {
-            let err =
-                crate::cli::actions::fetch_actions_job_log_via_gh("akiojin", "gwt", repo_path, 91)
-                    .expect_err("zip");
-            assert_eq!(err.kind(), io::ErrorKind::InvalidData);
-            assert!(err.to_string().contains("zip archive"));
-        });
-    }
-
-    #[test]
-    fn gh_wrappers_tolerate_resolve_failure_after_remote_state_changes() {
-        with_fake_gh("resolve-fails-but-resolved", |_repo_path| {
-            let resolved = crate::cli::pr::reply_and_resolve_pr_review_threads_via_gh(
-                "akiojin", "gwt", 12, "done",
-            )
-            .expect("resolved after retry");
-            assert_eq!(resolved, 2);
-        });
     }
 
     /// SPEC-1942 family split (FR-088〜092 / SC-025〜027): the parent
